requirements.txt
gunicorn==20.0.4
Flask==1.1.2
psycopg2-binary==2.8.6